Recorder: Expose frame number to client.
This change is adding meta frame number to BufferDescriptor,
this frame number is needed to match exact CameraMetaData
from camera result callback, currently it is matched using
timestamp which is not precise because meta buffer timestamp
and buffer timestamps may applied at different time in
camera backend.

this commit also fixing bug in adaptor, changing frame number
type from 64 bit to 32 bit as HAL accepts only 32 bit.

+++ common/qmmf_common_utils.h
@@ -54,8 +54,8 @@ const int64_t kWaitDelay = 2000000000;  // 2 sec
 
 struct StreamBuffer {
   CameraBufferMetaData info;
-  int64_t timestamp;
-  int64_t frame_number;
+  int64_t  timestamp;
+  uint32_t frame_number;
   uint32_t camera_id;
   uint32_t stream_id;
   android_dataspace data_space;
